Fix review nits
